refactor(http): separate notion of "http operation" from "http instance"
- http operation = design. it is a description of HOW an http instance should be put together
- http instance = runtime. it is the actual request/response.
